refactoring of app factory in flowap/__init__.py module.
